+++ src/main/java/dev/nicotopia/aoc/graphlib/NodeDistancePair.java
@@ -1,4 +1,4 @@
 package dev.nicotopia.aoc.graphlib;
 
-public record NodeDistancePair<NodeType>(NodeType node, int distance) {
+public record NodeDistancePair<NodeType>(NodeType node, long distance) {
 }
